ST-16: Move ResetGameDialog to its own class
